requirements.txt
gymnasium
moviepy
matplotlib
numpy
pyyaml
pyglet
tensorboardX
torch>=1.9.1
wandb>=0.17.4
open_clip_torch
torchvision
